+++ docs/source/release-notes.rst
@@ -1,5 +1,3 @@
-# coding=utf-8
-
 Release Notes
 =============
 
@@ -8,7 +6,7 @@ Release 0.2.0
 
 Improved fuzz testing.
 
-* Class ``FuzzExecutor`` making fuzz tests of applications with data files easy.
+* Class ``FuzzExecutor`` makes fuzz testing of applications taking data files easy.
 
 
 Release 0.1.0